A little of refactoring - snake game
Making classes for snake and food

There are some bugs, but they're related to pyglet and my understanding
of batch.
It seems to me that batch works like a stack.
It's interesting but it's ok for now.
I need to learn how to make commits and when, maybe how to use branches.
Maybe have a blog of my progress.